+++ pydidas/gui/frames/workflow_test_frame.py
@@ -147,7 +147,7 @@ class WorkflowTestFrame(BaseFrame):
     ``image_selection`` Parameter equals "Use scan indices").
     """
 
-    menu_icon = "pydidas::frame_icon_workflow_test.png"
+    menu_icon = "pydidas::frame_icon_workflow_test"
     menu_title = "Test workflow"
     menu_entry = "Workflow processing/Test workflow"