+++ pydidas/widgets/windows/global_settings_window.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ to view and modify the global settings in a seperate Window.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -46,7 +46,7 @@ class _GlobalSettingsWindow(PydidasWindow):
     A window to modify global pydidas settings.
     """
 
-    menu_icon = "qta::mdi.application-cog"
+    menu_icon = "mdi::application-cog-outline"
     menu_title = "Global configuration"
     menu_entry = "Global configuration"